requirements.txt
amqp==5.2.0
asgiref==3.8.1
autoflake==2.3.1
billiard==4.2.0
black==24.8.0
celery==5.4.0
certifi==2024.7.4
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
diff-match-patch==20230430
distlib==0.3.8
Django==5.0
django-background-tasks==1.2.8
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-filter==24.3
django-import-export==4.1.1
django-jazzmin==3.0.0
django-timezone-field==7.0
django-vue-admin==0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==28.0.0
filelock==3.15.4
flake8==7.1.1
identify==2.6.0
idna==3.8
kombu==5.4.0
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.8.0
prompt_toolkit==3.0.47
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
requirements.py==1.0.1
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
xero_python==6.1.0
gunicorn
